Improve agent progress calculation
- Highlight bootstrap node when installing cluster
- Improve agent progress calculation by using correct progress stages
